+++ packages/pwa-kit-runtime/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v3.14.0-dev (Sep 26, 2025)
 - Replace aws-serverless-express with @h4ad/serverless-adapter [#3325](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3325)
+- Added hybrid cdn for routing [#3393] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3393)
 
 ## v3.13.0 (Sep 25, 2025)